+++ src/adapter/etl-adapter-parquet/tests/Flow/ETL/Adapter/Parquet/Tests/Integration/ParquetTest.php
@@ -6,21 +6,20 @@ namespace Flow\ETL\Adapter\Parquet\Tests\Integration;
 
 use function Flow\ETL\Adapter\Parquet\{from_parquet, to_parquet};
 use function Flow\ETL\DSL\data_frame;
-use function Flow\ETL\DSL\{df, from_array, json_schema, schema, str_schema};
+use function Flow\ETL\DSL\{config, from_array, json_schema, schema, str_schema};
+use function Flow\Filesystem\DSL\{path};
 use Flow\ETL\Tests\Double\FakeExtractor;
 use Flow\ETL\{Tests\FlowTestCase};
-use Flow\Parquet\ParquetFile\Compressions;
-use Flow\Parquet\Reader;
 use Ramsey\Uuid\Uuid;
 
 final class ParquetTest extends FlowTestCase
 {
-    public function test_writing_to_file() : void
+    public function test_writing_and_reading_into_parquet() : void
     {
-        $path = __DIR__ . '/var/file.snappy.parquet';
-        $this->removeFile($path);
+        $path = path('memory://var/file.snappy.parquet');
 
-        df()
+        $config = config();
+        data_frame($config)
             ->read(new FakeExtractor(10))
             ->drop('null', 'array', 'object', 'enum')
             ->write(to_parquet($path))
@@ -28,28 +27,17 @@ final class ParquetTest extends FlowTestCase
 
         self::assertEquals(
             10,
-            (data_frame())
+            (data_frame($config))
                 ->read(from_parquet($path))
                 ->count()
         );
-
-        $parquetFile = (new Reader())->read($path);
-        self::assertNotEmpty($parquetFile->metadata()->columnChunks());
-
-        foreach ($parquetFile->metadata()->columnChunks() as $columnChunk) {
-            self::assertSame(Compressions::SNAPPY, $columnChunk->codec());
-        }
-
-        self::assertFileExists($path);
-        $this->removeFile($path);
     }
 
     public function test_writing_with_provided_schema() : void
     {
-        $path = __DIR__ . '/var/file_schema.snappy.parquet';
-        $this->removeFile($path);
-
-        df()
+        $path = path('memory://var/file_schema.snappy.parquet');
+        $config = config();
+        data_frame($config)
             ->read(from_array([
                 ['id' => 1, 'name' => 'test', 'uuid' => Uuid::fromString('26fd21b0-6080-4d6c-bdb4-1214f1feffef'), 'json' => '[{"id":1,"name":"test"},{"id":2,"name":"test"}]'],
                 ['id' => 2, 'name' => 'test', 'uuid' => Uuid::fromString('26fd21b0-6080-4d6c-bdb4-1214f1feffef'), 'json' => '[{"id":1,"name":"test"},{"id":2,"name":"test"}]'],
@@ -69,14 +57,13 @@ final class ParquetTest extends FlowTestCase
                 ['id' => '1', 'name' => 'test', 'uuid' => new \Flow\ETL\PHP\Value\Uuid('26fd21b0-6080-4d6c-bdb4-1214f1feffef'), 'json' => [['id' => 1, 'name' => 'test'], ['id' => 2, 'name' => 'test']]],
                 ['id' => '2', 'name' => 'test', 'uuid' => new \Flow\ETL\PHP\Value\Uuid('26fd21b0-6080-4d6c-bdb4-1214f1feffef'), 'json' => [['id' => 1, 'name' => 'test'], ['id' => 2, 'name' => 'test']]],
             ],
-            df()
+            data_frame($config)
                 ->read(from_parquet($path))
                 ->fetch()
                 ->toArray()
         );
 
-        self::assertFileExists($path);
-        $this->removeFile($path);
+        self::assertTrue($config->fstab()->for($path)->status($path)?->isFile());
     }
 
     /**