+++ src/bridge/symfony/http-foundation/src/Flow/Bridge/Symfony/HttpFoundation/Output/JsonOutput.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace Flow\Bridge\Symfony\HttpFoundation\Output;
 
 use function Flow\ETL\Adapter\JSON\to_json;
-use function Flow\Filesystem\DSL\path_stdout;
+use function Flow\Filesystem\DSL\{path_memory, path_stdout};
 use Flow\Bridge\Symfony\HttpFoundation\Output;
 use Flow\ETL\Loader;
 
@@ -23,7 +23,15 @@ final readonly class JsonOutput implements Output
 
     }
 
-    public function loader() : Loader
+    public function memoryLoader(string $id) : Loader
+    {
+        return to_json(path_memory($id, ['stream' => 'temp']))
+            ->withFlags($this->flags)
+            ->withDateTimeFormat($this->dateTimeFormat)
+            ->withRowsInNewLines($this->putRowsInNewLines);
+    }
+
+    public function stdoutLoader() : Loader
     {
         return to_json(path_stdout(['stream' => 'output']))
             ->withFlags($this->flags)