+++ src/bridge/symfony/http-foundation/src/Flow/Bridge/Symfony/HttpFoundation/Output/ParquetOutput.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace Flow\Bridge\Symfony\HttpFoundation\Output;
 
 use function Flow\ETL\Adapter\Parquet\to_parquet;
-use function Flow\Filesystem\DSL\path_stdout;
+use function Flow\Filesystem\DSL\{path_memory, path_stdout};
 use Flow\Bridge\Symfony\HttpFoundation\Output;
 use Flow\ETL\Loader;
 use Flow\ETL\Row\Schema;
@@ -25,7 +25,23 @@ final readonly class ParquetOutput implements Output
     ) {
     }
 
-    public function loader() : Loader
+    public function memoryLoader(string $id) : Loader
+    {
+        $loader = to_parquet(path_memory($id, ['stream' => 'temp']))
+            ->withCompressions($this->compressions);
+
+        if ($this->options !== null) {
+            $loader->withOptions($this->options);
+        }
+
+        if ($this->schema !== null) {
+            $loader->withSchema($this->schema);
+        }
+
+        return $loader;
+    }
+
+    public function stdoutLoader() : Loader
     {
         $loader = to_parquet(path_stdout(['stream' => 'output']))
             ->withCompressions($this->compressions);