+++ src/bridge/symfony/http-foundation/src/Flow/Bridge/Symfony/HttpFoundation/Response/FlowStreamedResponse.php
@@ -2,9 +2,10 @@
 
 declare(strict_types=1);
 
-namespace Flow\Bridge\Symfony\HttpFoundation;
+namespace Flow\Bridge\Symfony\HttpFoundation\Response;
 
 use function Flow\ETL\DSL\df;
+use Flow\Bridge\Symfony\HttpFoundation\Output;
 use Flow\ETL\{Config, Extractor, Transformation};
 use Flow\ETL\{Transformations};
 use Symfony\Component\HttpFoundation\StreamedResponse;
@@ -36,7 +37,7 @@ class FlowStreamedResponse extends StreamedResponse
             ->read($this->extractor)
             ->with($this->transformations)
             ->dropPartitions()
-            ->write($this->output->loader())
+            ->write($this->output->stdoutLoader())
             ->run();
     }
 }